feat: Adapt algorithms for measured dataset
